Update window_x11.go
fix build error

+++ gdk/window_x11.go
@@ -38,7 +38,7 @@ func (v *Window) GetXID() uint32 {
 
 //ForeignNewForDisplay is a wrapper around gdk_x11_window_foreign_new_for_display()
 // It only works on GDK versions compiled with X11 support - its return value can't be used if WorkspaceControlSupported returns false
-func (v *Display) ForeignNewForDisplay(xid uint64) (*Window, error) {
+func (v *Display) ForeignNewForDisplay(xid uint32) (*Window, error) {
 	c := C.gdk_x11_window_foreign_new_for_display(v.native(), C.guint64(xid))
 	if c == nil {
 		return nil, nilPtrErr
